% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_av}
\alias{asha_av}
\title{Indicador AV}
\usage{
asha_av(df, id, segundos, pop, raio)
}
\arguments{
\item{df}{Um dataframe contendo as variaveis necessarias para o calculo. Veja os detalhes.}

\item{id}{Variavel com o id das areas de saúde}

\item{segundos}{Variavel com o tempo de deslocamento em segundos}

\item{pop}{Variavel com o total de habitantes da
menor unidade territorial}

\item{raio}{Variavel numerica indicando o tamanho do raio (em minutos)}
}
\value{
A funcao adiciona 3 colunas: \code{minutos} (com os minutos de deslocamento),
        \code{av} (variavel binaria indicando se o setor esta no raio de acessibilidade viavel
        ou nao) e \code{av_prop} (que indica a proporcao da populacao no raio de acesso viavel
        em cada UBS.
}
\description{
Cria o indicador de acessibilidade viavel (AV)
}
\details{
Utiliza as variaveis passadas na funcao para calcular o indicador de acessibilidade
         viavel, incluindo a criacao das variaveis de acessibilidade.
}
\examples{
# asha_av(df, id, tempo, pop, raio)

}
\author{
Bruno Pinheiro
}
