% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{asha_bar}
\alias{asha_bar}
\title{Grafico de barras para uma variavel categorica}
\usage{
asha_bar(df, x, fill = NULL)
}
\arguments{
\item{df}{Um data.frame}

\item{x}{Uma variavel categorica}

\item{fill}{Uma variavel categorica}
}
\value{
Um grafico de barras para a variavel \code{x} da base \code{df}.
}
\description{
Uma funcao que pega uma variavel categorica, contabiliza suas classes
    e plota um grafico de barras formatado. O grafico sera plotado com coord_flip()
}
\details{
A funcao usa o dplyr para calcular e o ggplot para plotar.
    O parametro \code{x} deve ser uma string e portanto precisar ser
    incluido entre aspas, conforme o exemplo.
}
\examples{
data("setores_sp")
asha_bar(as.data.frame(setores_sp), "tipo")
# Com mais de uma variavel
library(dplyr)
setores_sp$pm <- as.factor(setores_sp$pessoas_setor > mean(setores_sp$pessoas_setor, na.rm = TRUE))
vars <- c("tipo", "pm")
plots <- lapply(vars, function(i) asha_bar(as.data.frame(setores_sp), i))

}
\seealso{
\code{\link[ggplot2]{geom_bar}},
         \code{\link[gridExtra]{grid.arrange}},
         \code{\link[scales]{percent}}
}
\author{
Bruno Pinheiro
}
