% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_dists}
\alias{asha_dists}
\title{Encontrar os n pontos mais proximos}
\usage{
asha_dists(fluxo, zonas, api)
}
\arguments{
\item{fluxo}{Uma matriz ou data frame de duas colunas representando a
latitude e a longitude das origens.}

\item{zonas}{Uma matriz ou data frame de duas colunas representando a
latitude e a longitude dos destinos.}

\item{api}{API do Google Distance Matrix API}
}
\value{
Retorna um data frame com as colunas \code{id2} (codigo de origem),
        \code{id1} (codigo de destino), \code{proximidade} e \code{distancia}.
}
\description{
A funcao busca e encontra o n pontos mais proximos
             entre dois datasets sf.
}
\details{
A funcao relaciona dois conjuntos de pontos espaciais e identifica os \code{n}
         pontos de \code{sf1} mais proximos de \code{sf2}. Usa a funcao nabor::knn
         para construir a matriz de distancia, filtra os \code{n} pontos e atribui os
         codigos de identificacao do destino e da origem.
}
\examples{
data("ubs_sp")
data("centroides_sp")
asha_nn(ubs_sp, centroides_sp, "cnes", "cd_geocodi", 3)

}
\seealso{
\code{\link[nabor]{knn}}
}
\author{
Bruno Pinheiro
}
