% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{asha_hist}
\alias{asha_hist}
\title{Histograma para uma variavel numerica}
\usage{
asha_hist(df, x, fill = NULL)
}
\arguments{
\item{df}{Um data.frame}

\item{x}{Uma variavel numerica}

\item{fill}{Um variavel categorica}
}
\value{
Um bonito histograma para a variavel \code{x} da base \code{df}.
}
\description{
Uma funcao que pega uma variavel numerica e plota
    um histograma. O grafico sera plotado ja formatado, com titulo
    chamando o nome da variavel.
}
\details{
A função usa o ggplot para plotar.
    O parametro \code{x} deve ser uma string e portanto precisar ser
    incluido entre aspas, conforme o exemplo.
}
\examples{
data("setores_sp")
asha_hist(as.data.frame(setores_sp), "pessoas_setor")

}
\seealso{
\code{\link[ggplot2]{geom_bar}}, \code{\link[gridExtra]{grid.arrange}}
}
\author{
Bruno Pinheiro
}
