% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_ind_group}
\alias{asha_ind_group}
\title{Calcular indicadores agregados}
\usage{
asha_ind_group(df, ind, pop, group, na.rm = FALSE)
}
\arguments{
\item{df}{Um dataframe contendo as variaveis necessarias para o calculo. Veja os detalhes.}

\item{ind}{Variavel categórica com o indicador}

\item{pop}{Variavel populacional}

\item{group}{Variável para agregar, além do indicador (opcional)}

\item{na.rm}{Útil caso a variável passada em \code{ind} contenha NA (default is FALSE)}
}
\value{
Descrever
}
\description{
Calcula população agregada por categorias
}
\details{
Descrever
}
\examples{
\dontrun{
asha_ind_group(od_indicadores, ac_cat, habitantes, malha, na.rm = TRUE)
}

}
\author{
Bruno Pinheiro
}
