% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_intersect}
\alias{asha_intersect}
\title{Interseccao de pontos de origem com areas de saude}
\usage{
asha_intersect(sf1, sf2, id1, id2)
}
\arguments{
\item{sf1}{Um objeto sf com geometria de poligonos}

\item{sf2}{Um objeto sf com geometria de pontos representando a origem da demanda}

\item{id1}{Codigo de identificacao dos poligonos}

\item{id2}{Codigo de identificao dos pontos}
}
\value{
Um data frame com \code{id1} (origem) e \code{id2} (area de cobertura) como
        colunas.
}
\description{
A funcao pega os pontos de origem e realiza interseccao
             com poligonos que demarcam territorios de saude
}
\details{
A funcao identifica os pontos localizados dentro de cada poligono.
         Usa a funcao sf::st_join para mesclar as bases com
         sf::st_intersect para realizar a intersecao.
}
\examples{
data("ubs_sp_areas")
data("centroides_sp")
asha_intersect(ubs_sp_areas, centroides_sp, "cnes", "cd_geocodi")

}
\author{
Bruno Pinheiro

#' @seealso \code{\link[sf]{st_intersects}},
         \code{\link[sf]{st_join}}
}
