% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{asha_map}
\alias{asha_map}
\title{Mapa cloropletico}
\usage{
asha_map(sf, x = NULL, pal = NULL)
}
\arguments{
\item{sf}{Um objeto espacial de classe \code{sf}}

\item{x}{Variavel categorica a ser usada em \code{group_by} e \code{fill}.
Parametro opcional, por padrao assume NULL.}

\item{pal}{Nome de uma paleta Color Brewer. Parametro opcional, por padrao
assume NULL: se \code{x} e dado, por padrao usa a paleta "Set2" e
se \code{x} nao e dado, nao usa nenhuma paleta.}
}
\value{
Um bonito histograma para a variavel \code{x} da base \code{df}.
}
\description{
Um tema simples para plotar mapas.
}
\details{
O tema cria um template limpo para o mapa, sem texto de eixo
         grid, bordas etc.
}
\examples{
data("ubs_sp")
asha_map(ubs_sp)

}
\seealso{
\code{\link[ggplot2]{theme}}
}
\author{
Bruno Pinheiro
}
