% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_nearest}
\alias{asha_nearest}
\title{Encontrar os n pontos mais proximos}
\usage{
asha_nearest(sf1, sf2, id1, id2, n)
}
\arguments{
\item{sf1}{Um objeto sf com geometria de pontos representando o destino}

\item{sf2}{Um objeto sf com geometria de pontos representando a origem}

\item{id1}{Codigo de identificacao do ponto de destino}

\item{id2}{Codigo de identificao da ponto de origem}

\item{n}{Numero de pontos mais proximo}
}
\value{
Retorna um data frame com as colunas \code{id_1} (destino) e \code{id_2} (origem),
        alem de uma coluna para cada \code{n} distancia.
}
\description{
A funcao busca e encontra o n pontos mais proximos
             entre dois datasets sf.
}
\details{
A funcao relaciona dois conjuntos de pontos espaciais e identifica os \code{n}
         pontos de \code{sf1} mais proximos de \code{sf2}. Usa a funcao nabor::knn
         para construir a matriz de distancia, filtra os \code{n} pontos e atribui
         codigos de identificacao.
}
\examples{
data("ubs_sp")
data("centroides_sp")
asha_nearest(ubs_sp, centroides_sp, "cnes", "cd_geocodi", 3)

}
\seealso{
\code{\link[nabor]{knn}}
}
\author{
Bruno Pinheiro
}
