% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_nn}
\alias{asha_nn}
\title{Encontrar os n pontos mais proximos}
\usage{
asha_nn(sf1, sf2, id1, id2, n)
}
\arguments{
\item{sf1}{Um objeto sf com geometria de pontos representando o destino}

\item{sf2}{Um objeto sf com geometria de pontos representando a origem}

\item{id1}{Codigo de identificacao do ponto de destino}

\item{id2}{Codigo de identificao do ponto de origem}

\item{n}{Numero de pontos mais proximo}
}
\value{
Retorna um data frame com as colunas \code{id2} (codigo de origem),
        \code{id1} (codigo de destino), \code{proximidade} e \code{distancia}.
}
\description{
A funcao busca e encontra o n pontos mais proximos
             entre dois datasets sf.
}
\details{
A funcao relaciona dois conjuntos de pontos espaciais e identifica os \code{n}
         pontos de \code{sf1} mais proximos de \code{sf2}. Usa a funcao nabor::knn
         para construir a matriz de distancia, filtra os \code{n} pontos e atribui os
         codigos de identificacao do destino e da origem.
}
\examples{
\dontrun{
data("ubs_pontos")
data("setores")
cent <- sf::st_centroid(setores)
asha_nn(ubs_pontos, cent, "cnes", "cd_geocodi", 3)
}

}
\seealso{
\code{\link[nabor]{knn}}
}
\author{
Bruno Pinheiro
}
