% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{asha_zones}
\alias{asha_zones}
\title{Estrutura a base de zonas necessaria para usar \code{asha_dists}}
\usage{
asha_zones(sf1, sf2, id1, id2)
}
\arguments{
\item{sf1}{Um objeto sf com coluna de id (por exemplo pontos de origem)}

\item{sf2}{Um objeto sf com coluna de id (por exemplo pontos de destino)}

\item{id1}{String especificando a variavel de id do objeto sf1}

\item{id2}{String especificando a variavel de id do objeto sf2}
}
\value{
Retorna um objeto sf com a coluna \code{id}, contendo os codigos de
        identificacao de sf1 e sf2, e a coluna \code{tipo}, referenciando os
        ids a partir dos nomes das variaves \code{id1} e \code{id2} passadas
        na funcao.
}
\description{
Mescla facilmente dois objetos sf em um
}
\details{
A funcao pega dois objetos espaciais de classe sf, mescla e
         organiza seus codigos id em uma unica coluna. E importante
         que os dois objetos tenham o mesmo crs definido previamente.
}
\examples{
zonas <- asha_zones(sf::st_centroid(setores), ubs_pontos, "cd_geocodi", "cnes")
str(zonas)

}
\author{
Bruno Pinheiro
}
