% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_saude_setores}
\alias{base_saude_setores}
\title{Pontos de UBS e centroides de setores censitarios unidos}
\format{Um dataset classe sf com pontos, contendo
        19409 linhas e 1 variavel:

\describe{
  \item{id}{Codigo de identificacao das origens e destinos}
}}
\usage{
base_saude_setores
}
\description{
O dataset reune os pontos de UBS e dos centroides dos setores
censitarios. É o formato padrão do parametro \code{zones} da
funcao \code{\link{asha_dists}}
}
\keyword{datasets}
