% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroides_sp}
\alias{centroides_sp}
\title{Centroides dos setores censitarios da cidade de Sao Paulo}
\format{Um dataset classe sf com os centroides dos setores censitarios,
    contendo 18953 linhas e 1 variavel
\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
}}
\usage{
centroides_sp
}
\description{
Criado com a funcao st_centroid a partir do dataset \link{setores_sp}.
}
\keyword{datasets}
