% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{get_censo}
\alias{get_censo}
\title{Get tabular data from IBGE Census 2010}
\usage{
get_censo(state, datatype = "tabular", mesh = "setores",
  savedir = NULL)
}
\arguments{
\item{state}{a string defining the the federation unitie to download data}

\item{datatype}{a string defining the kind of data to download: "tabular"
(default) or "geo"}

\item{mesh}{a string defining the mesh level to download: "setores"
(default) or "municipios"}

\item{savedir}{a string defining the directory to save the data. It's not
mandatory. See details.}
}
\value{
IBGE census tabular or geo data
}
\description{
Download and extract census data from IBGE
}
\details{
\code{get_censo} will download and extract IBGE data, keeping files
         in directory informed on \code{savedir} argument. If save dir is not
         informed, will store them in the temporary folder
}
\examples{
\dontrun{
get_censo("rr")
clean_tmp()
}

}
\author{
Bruno Pinheiro
}
