% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_functions.R
\name{get_zip}
\alias{get_zip}
\title{Get tabular data from IBGE}
\usage{
get_zip(url, savedir = NULL, junkpaths = TRUE)
}
\arguments{
\item{url}{a string defining url of zipped shapfile to get}

\item{savedir}{a string defining the directory to save the data. It's not
mandatory. See details.}

\item{junkpaths}{\code{unzip} TRUE/FALSE option to extract all files
directily and ignore folders. It's usefull for IBGE data, which
ussually have many junkpaths.}
}
\value{
extrated files
}
\description{
Download and extract tabular data from ftp://ftp.ibge.gov.br/
}
\details{
\code{get_zip} will download and extract files
}
\examples{
\dontrun{
get_zip()
}

}
\author{
Bruno Pinheiro
}
