% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od}
\alias{od}
\title{Setores censitarios da cidade de Sao Paulo}
\format{Um dataset classe sf com os poligonos dos setores censitarios,
    contendo 18953 linhas e 6 variaveis:
\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{cnes}{Codigo CNES da UBS}
  \item{dens_demografica}{Densidade demografica do setor censitario}
  \item{pessoas_sp}{Numero de pessoas residentes nos setores censitarios}
  \item{distancias}{Distancia em metros do percurso a pe entre o setor
  censitario e a UBS}
  \item{tempo}{Tempo de viagem do percurso a pe entre o setor censitario
  e a UBS em segundos}
  \item{total_enf}{Numero de enfermeiros nas UBS}
  \item{total_med}{Numero de medicos nas UBS}
  \item{oportunidades}{Indicador de Oportunidades (numero de medicos e enfermeiros
  nas UBS)}
  \item{modelo}{Variavel binaria indicando se a observacao se refere ao modelo
  vigente ou de proximidade}
  \item{demanda}{Indicador de Demanda das UBS}
  \item{ac}{Indicador de Acessibilidade Competitiva (AC)}
  \item{minutos}{Tempo de viagem do percurso a pe entre o setor censitario
  e a UBS em minutos}
  \item{av}{Variavel binaria indicando se o setor esta ou nao no raio de
  acessibilidade viavel por UBS}
  \item{av_prop_}{Proporcao de habitantes residentes no raio de
  acessibilidade viavel por UBS}
  \item{minutos_classe_}{Tempo de viagem do percurso a pe dividido em seis
  categorias}
  \item{av_prop_decimais_vig}{Decimos de proporcao de habitantes residentes
  no raio de acessibilidade viavel por UBS}
  \item{ac_classes}{Quintis de acessibilidade competitiva}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
od
}
\description{
O dataset foi estruturado a partir de dados do Censo 2010 disponibilidos
pelo IBGE (malha de setores censitarios do estado de Sao Paulo) e resultados
agregados por setor censitario para a cidade de Sao Paulo. Alem disso inclui
variaveis calcularas: o numero de habitantes, a area e densidade demografica.
As variaveis com sufixo _vig e _prox referem-se aos modelos vigente e de
proximidade, respectivamente.
}
\keyword{datasets}
