% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pessoas_sp}
\alias{pessoas_sp}
\title{Total de moradores nos setores censitários de São Paulo (SP)}
\format{Um dataframe com 18363 linhas e 2 variaveis:
\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{pessoas_setor}{Numero de pessoas residentes nos setores censitarios}
}}
\usage{
pessoas_sp
}
\description{
O dataset inclui uma variavel demografica extraida dos resultados do censo
2010 agregados por setor censitario contendo o numero de moradores em
domicilios particulares e coletivos.
}
\keyword{datasets}
