% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{populacao}
\alias{populacao}
\title{Número de habitantes por setor censitario}
\format{Um dataset classe \code{tibble()} contendo 18363 linhas
2 variaveis:

\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{habitantes}{Numero de habitantes em domicílios particulares
  e coletivos}
}}
\usage{
populacao
}
\description{
Sao os dados da variavel V001 do arquivo Domicilio02_SP1.csv dos
resultados do universo do Censo 2010 (IBGE), agregados por setor
censitario.
}
\keyword{datasets}
