% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prof_saude}
\alias{prof_saude}
\title{Profissionais de saude em Sao Paulo}
\format{Um dataset classe \code{tibble} contendo 18363 linhas
2 variaveis:
\describe{
  \item{cnes}{Codigo CNES da UBS}
  \item{enfermeiros}{Quantidade de enfermeiros}
  \item{medicos}{Quantidade de medicos}
}}
\usage{
prof_saude
}
\description{
Dados obtidos no DATASUS, com o total de medicos e enfermeiros nas
unidades de saude do municipio de Sao Paulo.
}
\keyword{datasets}
