% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotas_faltas}
\alias{rotas_faltas}
\alias{rotas_faltas}
\title{Pontos de UBS e centroides de setores censitarios unidos}
\format{Um data frame contendo 94765 linhas e 10 variaveis:

\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{cnes}{Codigo de identificacao da UBS}
  \item{ox}{Latitude da origem}
  \item{oy}{Longitude da origem}
  \item{dx}{Latitude do destino}
  \item{dy}{Longitude do destino}
  \item{de}{Endereco de origem}
  \item{para}{Endereco de destino}
  \item{distancias}{Distancia de caminhada em metros}
  \item{tempo}{Tempo de caminhada em segundos}
}}
\usage{
rotas_faltas

rotas_faltas
}
\description{
E a versao de ubs_sp_mobilidade com os casos que nao apresentaram
dados de rotas para modelo vigente entre as 5 UBS com menores
distancias euclidianas ate os setores censitarios.

O dataset reune os pontos de UBS e dos centroides dos setores
censitarios. É o formato padrão do parametro \code{zones} da
funcao \code{asha_dists}
}
\keyword{datasets}
