% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setores}
\alias{setores}
\title{Setores censitarios da cidade de Sao Paulo}
\format{Um dataset classe sf com os poligonos dos setores censitarios,
    contendo 18953 linhas e 5 variaveis:
\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{tipo}{Indica se o setor e urbano ou rural}
  \item{cd_geocodd}{Codigo de identificacao do distrito}
  \item{nm_distrit}{Nome do distrito}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
setores
}
\description{
O dataset foi estruturado a partir de dados do Censo 2010 disponibilidos
pelo IBGE (malha de setores censitarios do estado de Sao Paulo) e resultados
agregados por setor censitario para a cidade de Sao Paulo. Alem disso inclui
variaveis calcularas: o numero de habitantes, a area e densidade demografica.
As variaveis com sufixo _vig e _prox referem-se aos modelos vigente e de
proximidade, respectivamente.
}
\keyword{datasets}
