% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{theme_map}
\alias{theme_map}
\title{Tema para mapas}
\usage{
theme_map()
}
\description{
Um tema para mapas do ggplot2
}
\examples{
library(ggplot2)
data("ubs_sp")

# Map in geographic coordinates
ggplot(ubs_sp) +
    geom_sf() +
    theme_map()

}
