% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_malha}
\alias{ubs_malha}
\title{Areas de cobertura das UBS da cidade de Sao Paulo}
\format{Um data frame com 94765 linhas e 10 colunas
\describe{
  \item{cnes}{Codigo de identificacao da UBS}
  \item{nomeubs}{Nome da UBS}
  \item{sts}{Superintendencia de saude responsavel pela unidade}
  \item{crs}{Coordenadoria regional de saude responsavel pela unidade}
  \item{subpref}{Prefeitura regional onde a UBS esta localizada}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
ubs_malha
}
\description{
O dataset foi obtido da Secretaria Municipal de Saude por meio
de pedido de acesso a informacao. Traz os polígnos dos territórios
das Unidades Basicas de Saude, identificadas pelo número CNES de cada
equipamento. Foram adicionadas variaveis com a quantidade de medicos
e enfermeiros que trabalham em cada unidade.
}
\keyword{datasets}
