% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_pontos}
\alias{ubs_pontos}
\title{Localizacao das UBS da cidade de Sao Paulo}
\format{Um dataset classe sf com geometria de pontos,
    contendo 456 linhas e 2 variaveis:

\describe{
  \item{cnes}{Codigo de identificacao da UBS}
  \item{geometry}{Geometria dos pontos}
}}
\usage{
ubs_pontos
}
\description{
Geolocalização das UBS, representadas por pontos.
}
\keyword{datasets}
