% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_sp}
\alias{ubs_sp}
\title{Localizacao das UBS da cidade de Sao Paulo}
\format{Um dataset classe sf com os poligonos dos setores censitarios,
    contendo 456 linhas e 8 variaveis:
\describe{
  \item{cnes}{Codigo de identificacao da UBS}
  \item{nomeubs}{Nome da UBS}
  \item{sts}{Superintendencia de saude responsavel pela unidade}
  \item{crs}{Coordenadoria regional de saude responsavel pela unidade}
  \item{subpref}{Prefeitura regional onde a UBS esta localizada}
  \item{total_med}{Quantidade de medicos da UBS}
  \item{total_enf}{Quantidade de enfermeiros da UBS}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
ubs_sp
}
\description{
Traz as mesmas variaveis da base ubs_sp_areas, porem para os pontos de
localizacao das UBS e nao os poligonos das respectivas areas. As UBS sao
identificadas pelo seu numero CNES. Foram adicionadas variaveis com a
quantidade de medicos e enfermeiros que trabalham em cada unidade.
}
\keyword{datasets}
