% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_sp_areas}
\alias{ubs_sp_areas}
\title{Poligonos das areas de cobertura das UBS da cidade de Sao Paulo}
\format{Um dataset classe sf com os poligonos dos setores censitarios,
    contendo 456 linhas e 8 variaveis:
\describe{
  \item{cnes}{Codigo de identificacao da UBS}
  \item{nomeubs}{Nome da UBS}
  \item{sts}{Superintendencia de saude responsavel pela unidade}
  \item{crs}{Coordenadoria regional de saude responsavel pela unidade}
  \item{subpref}{Prefeitura regional onde a UBS esta localizada}
  \item{total_med}{Quantidade de medicos da UBS}
  \item{total_enf}{Quantidade de enfermeiros da UBS}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
ubs_sp_areas
}
\description{
O dataset foi obtido da Secretaria Municipal de Saude por meio
de pedido de acesso a informacao. Traz os polígnos dos territórios
das Unidades Basicas de Saude, identificadas pelo número CNES de cada
equipamento. Foram adicionadas variaveis com a quantidade de medicos
e enfermeiros que trabalham em cada unidade.
}
\keyword{datasets}
