% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_sp_mobilidade}
\alias{ubs_sp_mobilidade}
\title{Distacias e tempos de viagem a pe entre setores censitarios e UBS}
\format{Um data frame contendo 94765 linhas e 10 variaveis:

\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{cnes}{Codigo de identificacao da UBS}
  \item{ox}{Latitude da origem}
  \item{oy}{Longitude da origem}
  \item{dx}{Latitude do destino}
  \item{dy}{Longitude do destino}
  \item{de}{Endereco de origem}
  \item{para}{Endereco de destino}
  \item{distancias}{Distancia de caminhada em metros}
  \item{tempo}{Tempo de caminhada em segundos}
}}
\usage{
ubs_sp_mobilidade
}
\description{
O dataset foi criado com a funcao \code{asha_dist} e contem dados obtidos
da à Google Distance Matrix API. Inclui variaveis de distancias e tempos de
viagem entre os centroides dos setores censitarios da cidade de Sao Paulo e
as 5 UBS mais proximas de cada um.
}
\keyword{datasets}
