% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ubs_sp_profissionais}
\alias{ubs_sp_profissionais}
\title{Total de medicos e enfermeiros nas UBS de Sao Paulo}
\format{Um data frame com 456 linhas e 3 variaveis:

\describe{
  \item{cnes}{Codigo de identificacao da UBS}
  \item{total_med}{Quantidade de medicos da UBS}
  \item{total_enf}{Quantidade de enfermeiros da UBS}
}}
\usage{
ubs_sp_profissionais
}
\description{
O dataset conta com informacoes a respeito do total de profissionais
de atencao basica de saude (medicos e enfermeiros) trabalhando em
cada uma das UBS da cidade de Sao Paulo.
}
\keyword{datasets}
