% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zonas}
\alias{zonas}
\title{Pontos de UBS e centroides de setores censitarios unidos}
\format{Um dataset classe sf com pontos, contendo
        19409 linhas e 1 variavel:

\describe{
  \item{id}{Codigo de identificacao das origens e destinos}
}}
\usage{
zonas
}
\description{
O dataset reune os pontos de UBS e dos centroides dos setores
censitarios. É o formato padrão do parametro \code{zones} da
funcao \code{asha_dists}
}
\keyword{datasets}
