% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asha.R
\docType{package}
\name{asha}
\alias{asha}
\alias{asha-package}
\title{asha: Um pacote para analise de acessibilidade espacial a saude}
\description{
O pacote asha oferece xxx categorias de funcoes para a analise de dados da saude.
Basta utilizar os dados estruturados conforme os exemplos e utilizar suas funcoes.
}
\section{Funcoes do pacote}{

As funcoes do asha
}

\section{Outras informacoes}{

Veja mais em \link{asha_bar}, \link{asha_hist}
}

\examples{
# plotar variavel numerica
asha_hist(ubs_sp_areas, "total_enf")
# plotar variavel categorica
asha_bar(as.data.frame(setores_sp), "tipo")

}
