% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setores_sp}
\alias{setores_sp}
\title{Poligonos de setores censitarios da cidade de Sao Paulo}
\format{Um dataset classe sf com os poligonos dos setores censitarios,
    contendo 18953 linhas e 6 variaveis:
\describe{
  \item{cd_geocodi}{Codigo de identificacao do setor censitario}
  \item{cd_geocodd}{Codigo de identificacao do distrito}
  \item{nm_distrit}{Nome do distrito}
  \item{tipo}{Indica se o setor e urbano ou rural}
  \item{pessoas_setor}{Numero de pessoas residentes nos setores censitarios}
  \item{area}{Area do setor censitario em km^2^}
  \item{dens_demografica}{Densidade demografica do setor censitario}
  \item{geometry}{Geometria dos polígonos}
}}
\usage{
setores_sp
}
\description{
O dataset foi estruturado a partir de dados do Censo 2010 disponibilidos
pelo IBGE (malha de setores censitarios do estado de Sao Paulo) e resultados
agregados por setor censitario para a cidade de Sao Paulo. Alem disso inclui
variaveis calcularas: o numero de habitantes, a area e densidade demografica.
}
\keyword{datasets}
